refactor int conversions

+++ replica/src/baxos.go
@@ -10,13 +10,13 @@ import (
 
 type BaxosProposerInstance struct {
 	preparedBallot        *common.Ballot // the ballot number for which the prepare message was sent
-	numSuccessfulPromises int32 // the number of successful promise messages received
+	numSuccessfulPromises int // the number of successful promise messages received
 
 	highestSeenAcceptedBallot *common.Ballot       // the highest accepted ballot number among them set of Promise messages
 	highestSeenAcceptedValue  *common.WriteRequest // the highest accepted value among the set of Promise messages
 
 	proposedValue        *common.WriteRequest // the value that is proposed
-	numSuccessfulAccepts int32               // the number of successful accept messages received
+	numSuccessfulAccepts int               // the number of successful accept messages received
 }
 
 type BaxosAcceptorInstance struct {
@@ -45,7 +45,7 @@ type ReadRequestInstance struct {
 */
 
 type Baxos struct {
-	lastCommittedLogIndex int32                   // the last log position that is committed
+	lastCommittedLogIndex int                   // the last log position that is committed
 	replicatedLog         []BaxosInstance         // the replicated log of commands
 	timer                 *common.TimerWithCancel // the timer for collecting promise / accept responses
 	roundTripTime         int64                   // network round trip time in microseconds
@@ -61,7 +61,7 @@ type Baxos struct {
 
 	isAsync      bool // to simulate an asynchronous network
 
-	quorumSize int32
+	quorumSize int
 }
 
 /*
@@ -90,7 +90,7 @@ func InitBaxosConsensus(replica *Replica, isAsync bool, roundTripTime int64) *Ba
 		retries:               0,
 		replica:               replica,
 		isAsync:               isAsync,
-		quorumSize:            int32(replica.numReplicas/2 + 1),
+		quorumSize:            int(replica.numReplicas/2 + 1),
 	}
 }
 
@@ -164,12 +164,12 @@ func (rp *Replica) createInstance(n int) {
 			proposer_bookkeeping: BaxosProposerInstance {
 				preparedBallot: &common.Ballot{
 					Number:    -1,
-					ReplicaId: rp.id,
+					ReplicaId: int64(rp.id),
 				},
 				numSuccessfulPromises:     0,
 				highestSeenAcceptedBallot: &common.Ballot{
 					Number:    -1,
-					ReplicaId: rp.id,
+					ReplicaId: int64(rp.id),
 				},
 				highestSeenAcceptedValue:  &common.WriteRequest{},
 				proposedValue:             &common.WriteRequest{},
@@ -178,11 +178,11 @@ func (rp *Replica) createInstance(n int) {
 			acceptor_bookkeeping: BaxosAcceptorInstance{
 				promisedBallot: &common.Ballot{
 					Number:    -1,
-					ReplicaId: rp.id,
+					ReplicaId: int64(rp.id),
 				},
 				acceptedBallot: &common.Ballot{
 					Number:    -1,
-					ReplicaId: rp.id,
+					ReplicaId: int64(rp.id),
 				},
 				acceptedValue:  &common.WriteRequest{},
 			},
@@ -203,7 +203,7 @@ func (rp *Replica) printBaxosLogConsensus() {
 	}
 	defer f.Close()
 
-	for i := int32(0); i <= rp.baxosConsensus.lastCommittedLogIndex; i++ {
+	for i := 0; i <= rp.baxosConsensus.lastCommittedLogIndex; i++ {
 		if !rp.baxosConsensus.replicatedLog[i].decided {
 			panic("should not happen")
 		}
@@ -220,7 +220,7 @@ func (rp *Replica) printBaxosLogConsensus() {
 
 func (rp *Replica) updateSMR() {
 
-	for i := rp.baxosConsensus.lastCommittedLogIndex + 1; i < int32(len(rp.baxosConsensus.replicatedLog)); i++ {
+	for i := rp.baxosConsensus.lastCommittedLogIndex + 1; i < len(rp.baxosConsensus.replicatedLog); i++ {
 
 		baxosInstance := rp.baxosConsensus.replicatedLog[i]
 
@@ -234,7 +234,7 @@ func (rp *Replica) updateSMR() {
 			Sender:  int64(rp.id),
 		}
 		rp.baxosConsensus.lastCommittedLogIndex = i
-		rp.sendClientResponse(&response, int32(baxosInstance.decidedValue.Sender))
+		rp.sendClientResponse(&response, int(baxosInstance.decidedValue.Sender))
 		rp.debug(fmt.Sprintf("Committed baxos consensus instance %d", i), 0)
 	}
 }
